modify: renamte userListsContainer
